+++ src/gbaEmu/Register.java
@@ -4,16 +4,16 @@ public class Register {
 	byte a,b,c,d,e,f,h,l; //int can cause problem because register is 8-bit.
 	boolean zf, nf, hf, cf;
 	short pc, sp;
-	public int af() {
-		return a << 8 | f;
+	public short af() {
+		return (short) (a << 8 | f);
 	}
-	public int hl() {
-		return h << 8 | l;
+	public short hl() {
+		return (short) (h << 8 | l);
 	}
-	public int bc() {
-		return b << 8 | c;
+	public short bc() {
+		return (short) (b << 8 | c);
 	}
-	public int de() {
-		return d << 8 | e;
+	public short de() {
+		return (short) (d << 8 | e);
 	}
 }
